# docs/swagger.yaml
definitions:
  be-dashboard-nba_internal_presentation_presenter.Pagination:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponseErrorPayload:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.Pagination'
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_menu.UpdateMenuOrderRequest:
    properties:
      group:
        type: string
      parent_id:
        minimum: 1
        type: integer
      sorted_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - group
    - sorted_ids
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_role.CreateRoleRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - code
    - name
    type: object
  be-dashboard-nba_internal_presentation_presenter_role.ReadRoleResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_role.UpdateRoleRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
  be-dashboard-nba_internal_presentation_presenter_user.CreateUserRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      role_id:
        type: integer
    required:
    - active
    - email
    - full_name
    - name
    - password
    - role_id
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_user.ReadUserResponse:
    properties:
      active:
        type: boolean
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: Template API for Golang project
  title: Base APP Go
  version: "1.0"
paths:
...
  /menus/reorder:
    put:
      consumes:
      - application/json
      description: Updates the sort order of menus using drag-and-drop. Requires Bearer
        token.
      parameters:
      - description: New menu order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.UpdateMenuOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update menu order
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid JSON or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update Menu Order (Reorder)
      tags:
      - Menu
...
  /roles:
    get:
      description: Retrieves a paginated and searchable list of all roles. Requires
        Bearer token.
      parameters:
      - description: Search term for role name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Sort order (e.g., name ASC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get roles
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_role.ReadRoleResponse'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Roles List
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Creates a new role. Requires Bearer token.
      parameters:
      - description: Role data to create
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_role.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid JSON or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponseErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed create role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Create Role
      tags:
      - Role
  /roles/{role_id}:
    delete:
      description: Deletes a role by its ID. Requires Bearer token.
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid Role ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized (Missing Auth)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Role ID not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed delete role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Delete Role
      tags:
      - Role
    get:
      description: Fetches the details of a single role by its ID. Requires Bearer
        token.
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get role detail
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_role.ReadRoleResponse'
              type: object
        "400":
          description: Bad Request (Invalid Role ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Role ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get role detail
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Role Detail
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Updates an existing role by its ID. Requires Bearer token.
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      - description: Role data to update
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_role.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponseErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Role ID not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed update role
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update Role
      tags:
      - Role
...
  /users:
    get:
      description: Retrieves a paginated and searchable list of users in the system.
        Requires Bearer token.
      parameters:
      - description: Search term for name or email
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Sort order (e.g., name ASC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get users
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.ReadUserResponse'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get users
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details. Requires Bearer token.
      parameters:
      - description: User data to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create user
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid JSON or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Role ID not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your token}" into the field below.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"